+++ src/conexaoBanco/ConexaoMySQL.java
@@ -1,0 +1,9 @@
+package conexaoBanco;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class ConexaoMySQL {
+    
+    
+}// class
