+++ src/conexaoBanco/create-database-bdhotel.sql
@@ -1,0 +1,10 @@
+create database IF NOT EXISTS bdhotel;
+use bdhotel;
+CREATE TABLE IF NOT EXISTS tblhotel (
+hotelid INTEGER AUTO_INCREMENT,
+nome VARCHAR(255),
+cidade VARCHAR(255),
+quartos INTEGER,
+valorDiaria DOUBLE,
+PRIMARY KEY (hotelid)
+);
